Post push task and install dependency bugfixes

+++ dataschema_manager/tasks.py
@@ -2,7 +2,7 @@ import json
 import logging
 
 from MOOCworkbench.celery import app
-from git_manager.helpers.helper import get_experiment_from_repo_name
+from git_manager.helpers.helper import get_exp_or_package_from_repo_name
 from experiments_manager.consumers import send_message
 from experiments_manager.helper import MessageStatus
 from experiments_manager.models import Experiment
@@ -30,19 +30,20 @@ def task_write_data_schema(experiment_id):
 
 @app.task
 def task_read_data_schema(repository_name):
-    experiment = get_experiment_from_repo_name(repository_name)
-    logger.debug('reading data schema for: %s', experiment)
-    github_helper = GitHubHelper(experiment.owner.user, experiment.git_repo.name)
-    schema_json = json.loads(github_helper.view_file('schema/schema.json'))
-    data_schema_fields = parse_json_table_schema(schema_json)
-    data_schema = experiment.schema
-    for field in data_schema.fields.all():
-        data_schema.fields.remove(field)
-        data_schema.save()
-        field.delete()
-    for new_field in data_schema_fields:
-        new_field.save()
-        data_schema.fields.add(new_field)
-        data_schema.save()
-    logger.debug('read data schema for: %s', experiment, data_schema_fields)
+    experiment = get_exp_or_package_from_repo_name(repository_name)
+    if isinstance(experiment, Experiment):
+        logger.debug('reading data schema for: %s', experiment)
+        github_helper = GitHubHelper(experiment.owner.user, experiment.git_repo.name)
+        schema_json = json.loads(github_helper.view_file('schema/schema.json'))
+        data_schema_fields = parse_json_table_schema(schema_json)
+        data_schema = experiment.schema
+        for field in data_schema.fields.all():
+            data_schema.fields.remove(field)
+            data_schema.save()
+            field.delete()
+        for new_field in data_schema_fields:
+            new_field.save()
+            data_schema.fields.add(new_field)
+            data_schema.save()
+        logger.debug('read data schema for: %s', experiment, data_schema_fields)
